add modules for individual classes
