+++ src/main.py
@@ -1,88 +1,7 @@
-import random
+from word import Words
+from user import Users
+from command import Command
 from settings import settings
-
-
-class User:
-    def __init__(self, name):
-        self.id = random.choice(range(1_000_000, 9_999_999))
-        self.name = name
-        self.used_words = []
-        self.score = 0
-        self.is_alive = True
-
-    def add_word(self, word):
-        self.used_words.append(word)
-        self.score += 1
-
-    def remove_word(self):
-        if len(self.used_words):
-            self.used_words.pop()
-            self.score -= 1
-
-
-class Users:
-    def __init__(self, names=[]):
-        self.users = [User(name) for name in names]
-
-        self.current_user_index = random.choice(range(self.user_count))
-        self.current_user = self.users[self.current_user_index]
-
-    def add_user(self, name):
-        self.users.append(User(name))
-
-    def kill_user(self, user):
-        for current_user in self.users:
-
-            if user.id == current_user.id:
-                current_user.is_alive = False
-                self.rotate_forward()
-
-                break
-
-    def rotate_forward(self):
-        if self.current_user_index == self.user_count - 1:
-            self.current_user_index = 0
-
-        else:
-            self.current_user_index += 1
-
-        self.current_user = self.users[self.current_user_index]
-
-        if not self.current_user.is_alive:
-            self.rotate_forward()
-
-    def rotate_backward(self):
-        if self.current_user_index == 0:
-            self.current_user_index = self.user_count - 1
-
-        else:
-            self.current_user_index -= 1
-
-        self.current_user = self.users[self.current_user_index]
-
-        if not self.current_user.is_alive:
-            self.rotate_backward()
-
-    @property
-    def only_one_alive(self):
-        alive_users = 0
-
-        for user in self.users:
-            if user.is_alive:
-                alive_users += 1
-
-                if alive_users > 1:
-                    return False
-
-        return True
-
-    @property
-    def user_count(self):
-        return len(self.users)
-
-    @property
-    def previous_user(self):
-        return self.users[-1]
 
 
 while True:
@@ -120,61 +39,6 @@ while True:
         break
 
 
-class Word:
-    def __init__(self, word, user):
-        self.word = word
-        self.last_letter = word[-1]
-        self.user = user
-
-
-class Words:
-    def __init__(self, settings):
-        self.used_words = []
-        self.used_words_ref = []
-
-        initial_word = settings.initial_word
-        self.used_words.append(Word(initial_word, User("base")))
-        self.used_words_ref.append(initial_word)
-
-    def add_word(self, word, user):
-        created_word = Word(word, user)
-
-        self.used_words.append(created_word)
-        self.used_words_ref.append(word)
-
-        user.add_word(created_word)
-
-    def remove_word(self):
-        removed_word = self.used_words.pop()
-        self.used_words_ref.pop()
-
-        return removed_word
-
-    def is_used(self, word):
-        return word in self.used_words_ref
-
-    @property
-    def last_word(self):
-        return self.used_words[-1]
-
-    @property
-    def can_revert(self):
-        if len(self.used_words_ref) > 1:
-            return True
-        else:
-            return False
-
-
-class Command:
-    def __init__(self, string, command_prefix):
-        self.is_command = string.startswith(command_prefix)
-
-        if self.is_command:
-            self.action = string[string.index(command_prefix) + 1:]
-        else:
-            self.action = None
-
-
 words = Words(settings)
 
 initial_word = settings.initial_word
@@ -199,8 +63,7 @@ while True:
 
     user_word_input = input(f"[{current_user.name}] :: ").strip().lower()
 
-    command_prefix = settings.command_prefix
-    command = Command(user_word_input, command_prefix)
+    command = Command(user_word_input)
 
     if command.is_command:
